fix: limit chain options on the form
