+++ frontend/app/src/types/notes.ts
@@ -27,6 +27,7 @@ export enum NoteLocation {
   ACCOUNTS_EVM = 'ACCOUNTS_EVM',
   ACCOUNTS_BITCOIN = 'ACCOUNTS_BITCOIN',
   ACCOUNTS_SUBSTRATE = 'ACCOUNTS_SUBSTRATE',
+  ACCOUNTS_VALIDATOR = 'ACCOUNTS_VALIDATOR',
   BALANCES_BLOCKCHAIN = 'ACCOUNTS_BALANCES_BLOCKCHAIN',
   BALANCES_EXCHANGE = 'ACCOUNTS_BALANCES_EXCHANGE',
   BALANCES_NON_FUNGIBLE = 'ACCOUNTS_BALANCES_NON_FUNGIBLE',
